Removing Authorization namespace
