+++ src/Microsoft.AspNet.Security/Microsoft.AspNet.Security.kproj
@@ -25,10 +25,12 @@
     <Compile Include="AuthenticationMode.cs" />
     <Compile Include="AuthenticationOptions.cs" />
     <Compile Include="AuthenticationTicket.cs" />
-    <Compile Include="Authorization\AuthorizationPolicyContext.cs" />
-    <Compile Include="Authorization\DefaultAuthorizationService.cs" />
-    <Compile Include="Authorization\IAuthorizationPolicy.cs" />
-    <Compile Include="Authorization\IAuthorizationService.cs" />
+    <Compile Include="AuthorizationPolicy.cs" />
+    <Compile Include="AuthorizationPolicyContext.cs" />
+    <Compile Include="AuthorizationServiceExtensions.cs" />
+    <Compile Include="DefaultAuthorizationService.cs" />
+    <Compile Include="IAuthorizationPolicy.cs" />
+    <Compile Include="IAuthorizationService.cs" />
     <Compile Include="CertificateSubjectKeyIdentifierValidator.cs" />
     <Compile Include="CertificateSubjectPublicKeyInfoValidator.cs" />
     <Compile Include="CertificateThumbprintValidator.cs" />
